Refactor for TypeScript Add tests for supported encodings
